chore(wallet): move HW connect TSX to a separate component
