+++ components/brave_wallet_ui/stories/wallet-concept.tsx
@@ -299,10 +299,6 @@ export const _DesktopWalletConcept = (args: { onboarding: boolean, locked: boole
     alert(`Account Name: ${name}, Private Key: ${key}`)
   }
 
-  const onConnectHardwareWallet = (hardware: 'Ledger' | 'Trezor') => {
-    alert(`Connecting to ${hardware} wallet`)
-  }
-
   const onToggleAddModal = () => {
     setShowAddModal(!showAddModal)
   }
@@ -464,7 +460,6 @@ export const _DesktopWalletConcept = (args: { onboarding: boolean, locked: boole
                           selectedUSDAssetPrice={selectedUSDAssetPrice}
                           selectedBTCAssetPrice={selectedBTCAssetPrice}
                           userAssetList={userAssetList}
-                          onConnectHardwareWallet={onConnectHardwareWallet}
                           onCreateAccount={onCreateAccount}
                           onImportAccount={onImportAccount}
                           isLoading={false}
